requirements.txt
numpy
pandas
pillow
opencv-python
matplotlib
psutil